cli: use sopel.cli.utils to load settings in run_script
